Render <SearchFilesPopover />
